Add functions to generalize processing
While there are only a few data sources to parse through, there could be
more report sources in the future. Refactoring out the repetitive code
that parses the csv files makes the code more readable and easier to
extend in the future. This is a first attempt at generalizing the code.